Extract CPP #define tables, too
The TCG TPM2 part 2 "Structures" and the TCG Algorithm Registry contain a couple
of tables containing CPP-#define-like content. Extract these as well.

Note that this includes the ECC curve parameter specification tables from the
Algorithm registry, which are in fact *not* #define-like, but these might be
useful to have later anyway.

Signed-off-by: Nicolai Stange <nstange@suse.de>